# docs/swagger.yaml
basePath: /
definitions:
...
  author.Author:
    properties:
      bio:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      linkedinUrl:
        type: string
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      xUrl:
        type: string
    type: object
...
  category.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.UpdateImageMetadataRequest:
    properties:
      is_active:
        type: boolean
      title:
        type: string
    type: object
...
  industry_news.IndustryNews:
    properties:
      author:
        $ref: '#/definitions/author.Author'
      authorId:
        type: integer
      category:
        $ref: '#/definitions/category.Category'
      categoryId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      internalLinks:
        items:
          $ref: '#/definitions/industry_news.InternalLinkEntry'
        type: array
      location:
        type: string
      metadata:
        $ref: '#/definitions/industry_news.IndustryNewsMetadata'
      publishDate:
        type: string
      reviewedAt:
        type: string
      reviewedBy:
        type: integer
      scheduledPublishEnabled:
        type: boolean
      slug:
        type: string
      status:
        $ref: '#/definitions/industry_news.IndustryNewsStatus'
      tags:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  industry_news.IndustryNewsImage:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      industry_news:
        $ref: '#/definitions/industry_news.IndustryNews'
      industry_news_id:
        type: integer
      is_active:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: integer
    type: object
...
  industry_news.IndustryNewsMetadata:
    properties:
      keywords:
        items:
          type: string
        type: array
      metaDescription:
        type: string
      metaTitle:
        type: string
    type: object
...
  industry_news.IndustryNewsStatus:
    enum:
    - draft
    - review
    - published
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusReview
    - StatusPublished
  industry_news.InternalLinkEntry:
    properties:
      keyword:
        type: string
      linkedCount:
        type: integer
      targetId:
        type: integer
      targetTitle:
        type: string
      targetType:
        description: '"report", "blog", "press-release", "industry-news"'
        type: string
      targetUrl:
        type: string
    type: object
...
  response.Meta:
    properties:
      cursor:
        type: string
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: 192.168.1.2:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: RESTful API for healthcare market research reports and categories with
    caching and pagination support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Healthcare Market Research API
  version: "1.0"
paths:
...
  /api/v1/industry-news/{newsId}/images:
    get:
      consumes:
      - application/json
      description: Get all images for an industry news article with optional active
        filter. Admin/editor only.
      parameters:
      - description: Industry news article ID
        in: path
        name: newsId
        required: true
        type: integer
      - description: Filter by active status (true = only active images)
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of images
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/industry_news.IndustryNewsImage'
                  type: array
              type: object
        "400":
          description: Bad request - invalid industry news ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: List industry news images
      tags:
      - Industry News Images
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for an industry news article (for use in the content
        editor). Admin/editor only.
      parameters:
      - description: Industry news article ID
        in: path
        name: newsId
        required: true
        type: integer
      - description: 'Image file (max 10MB, allowed types: JPEG, PNG, WebP, GIF)'
        in: formData
        name: image
        required: true
        type: file
      - description: Image title (optional, max 255 chars)
        in: formData
        name: title
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Image uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/industry_news.IndustryNewsImage'
              type: object
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Industry news article not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Upload industry news image
      tags:
      - Industry News Images
  /api/v1/industry-news/images/{imageId}:
    delete:
      consumes:
      - application/json
      description: Delete an industry news image. Admin/editor only.
      parameters:
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad request - invalid image ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Image not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete industry news image
      tags:
      - Industry News Images
    get:
      consumes:
      - application/json
      description: Get a single industry news image by ID. Admin/editor only.
      parameters:
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image details
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/industry_news.IndustryNewsImage'
              type: object
        "400":
          description: Bad request - invalid image ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Image not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get single industry news image
      tags:
      - Industry News Images
    patch:
      consumes:
      - application/json
      description: Update image title and/or is_active status. Supports partial updates.
        Admin/editor only.
      parameters:
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      - description: Metadata to update (all fields are optional for partial updates)
        in: body
        name: metadata
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateImageMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated image
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/industry_news.IndustryNewsImage'
              type: object
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Image not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update industry news image metadata
      tags:
      - Industry News Images
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...